pyproject.toml
[tool.poetry]
name = "autochapters"
version = "0.1.0"
description = "Automatically create video chapters from transcripts"
authors = ["Kobi Felton <kobi.c.f@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.55"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
